requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
dill==0.3.8
environs==9.5.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.60.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
mdurl==0.1.2
minio==7.2.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydeck==0.8.1b0
Pygments==2.17.2
pymilvus==2.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
validators==0.22.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
